refactor: remove nested logframe pg component and polish quick nav styles
